+++ semantic_router/schemas/route.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Route(BaseModel):
+    name: str
+    utterances: list[str]
+    description: str | None = None
