+++ tests/unit/encoders/test_tfidf.py
@@ -1,6 +1,6 @@
 import pytest
 from semantic_router.encoders import TfidfEncoder
-from semantic_router.schema import Route
+from semantic_router.schemas.route import Route
 
 
 @pytest.fixture
