+++ tests/unit/test_hybrid_layer.py
@@ -8,7 +8,7 @@ from semantic_router.encoders import (
     TfidfEncoder,
 )
 from semantic_router.hybrid_layer import HybridRouteLayer
-from semantic_router.schema import Route
+from semantic_router.schemas.route import Route
 
 
 def mock_encoder_call(utterances):
